move all cannotberenderedreason logic to base track, other tracks may override this (such as sequence track). track configs set their own maxdisplayedbpperpx, restore model.tsx->model.ts
