+++ packages/wiggle/src/WiggleTrack/model.ts
@@ -265,9 +265,18 @@ const stateModelFactory = (configSchema: ReturnType<typeof ConfigSchemaF>) =>
                 try {
                   const aborter = new AbortController()
                   self.setLoading(aborter)
-
                   const view = getContainingView(self) as LGV
-                  if (!view.initialized && !self.ready) {
+                  console.log(
+                    !view.initialized,
+                    !self.ready,
+                    view.bpPerPx,
+                    self.maxViewBpPerPx,
+                  )
+                  if (
+                    (!view.initialized && !self.ready) ||
+                    view.bpPerPx > self.maxViewBpPerPx
+                  ) {
+                    console.log('in new return statement')
                     return
                   }
